Moved equality descriptions to new utils class.
Added new relative equality predicates for integer numbers.
