+++ gdsc-test-api/src/main/java/uk/ac/sussex/gdsc/test/api/comparisons/EqualityUtils.java
@@ -31,16 +31,6 @@ import java.math.BigInteger;
  */
 public final class EqualityUtils {
 
-  /** Constant used to ignore the relative error. */
-  private static final double IGNORE_RELATIVE_ERROR = -1;
-  /** Constant used to ignore the double absolute error. */
-  private static final double IGNORE_DOUBLE_ABSOLUTE_ERROR = -1;
-  /** Constant used to ignore the float absolute error. */
-  private static final float IGNORE_FLOAT_ABSOLUTE_ERROR = -1;
-  /** Constant used to ignore the long absolute error. */
-  private static final long IGNORE_LONG_ABSOLUTE_ERROR = -1;
-  /** Constant used to ignore the int absolute error. */
-  private static final int IGNORE_INT_ABSOLUTE_ERROR = -1;
   /** The maximum symmetric relative error. */
   private static final double MAX_RELATIVE_ERROR = 2;
   /**
@@ -60,22 +50,6 @@ public final class EqualityUtils {
    * The maximum absolute error between two {@code byte} values.
    */
   public static final int MAX_BYTE_ABS_ERROR = 0xff;
-  /**
-   * The description of symmetric relative error {@code <=}.
-   */
-  private static final String DESCRIPTION_REL_ERROR_LTE = "|v1-v2|/max(|v1|,|v2|) <= ";
-  /**
-   * The description of asymmetric relative error {@code <=}.
-   */
-  private static final String DESCRIPTION_ASYM_REL_ERROR_LTE = "|v1-v2|/|v1| <= ";
-  /**
-   * The description of absolute error {@code <=}.
-   */
-  private static final String DESCRIPTION_ABS_ERROR_LTE = "|v1-v2| <= ";
-  /**
-   * The description of absolute error {@code == 0}.
-   */
-  private static final String DESCRIPTION_ABS_ERROR_0 = "|v1-v2| == 0";
 
   /**
    * Do not allow public construction.
@@ -491,107 +465,6 @@ public final class EqualityUtils {
     // byte uses int arithmetic by default
     final int delta = (value1 > value2) ? value1 - value2 : value2 - value1;
     return (delta <= absoluteError);
-  }
-
-  /**
-   * Get the description of the test that two doubles are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with
-   * {@link #doublesValidateAbsoluteError(double)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String doublesGetDescriptionWithin(double absoluteError) {
-    return (absoluteError == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
-  }
-
-  /**
-   * Get the description of the test that two floats are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with
-   * {@link #floatsValidateAbsoluteError(float)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String floatsGetDescriptionWithin(float absoluteError) {
-    return (absoluteError == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
-  }
-
-  /**
-   * Get the description of the test that two longs are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with
-   * {@link #longsValidateAbsoluteError(BigInteger)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String longsGetDescriptionWithin(BigInteger absoluteError) {
-    return (absoluteError.signum() == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
-  }
-
-  /**
-   * Get the description of the test that two longs are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with {@link #longsValidateAbsoluteError(long)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String longsGetDescriptionWithin(long absoluteError) {
-    return (absoluteError == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
-  }
-
-  /**
-   * Get the description of the test that two ints are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with {@link #intsValidateAbsoluteError(long)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String intsGetDescriptionWithin(long absoluteError) {
-    return (absoluteError == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
-  }
-
-  /**
-   * Get the description of the test that two shorts are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with {@link #shortsValidateAbsoluteError(int)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String shortsGetDescriptionWithin(int absoluteError) {
-    return (absoluteError == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
-  }
-
-  /**
-   * Get the description of the test that two bytes are equal within an absolute error.
-   *
-   * <p>It is assumed the errors have been validated with {@link #bytesValidateAbsoluteError(int)}.
-   *
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  static String bytesGetDescriptionWithin(int absoluteError) {
-    return (absoluteError == 0) ? DESCRIPTION_ABS_ERROR_0
-        : DESCRIPTION_ABS_ERROR_LTE + absoluteError;
   }
 
   ////////////////////////////////////////////////////////////////////////
@@ -1145,287 +1018,6 @@ public final class EqualityUtils {
     return (value <= 0) ? value : -value;
   }
 
-  //@formatter:off
-  /**
-   * Get the description of the test that two values are close using a relative and
-   * absolute error. The relative error is symmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum relative error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String doublesGetDescriptionClose(double relativeError, double absoluteError) {
-    return doubleGetDescriptionAreClose(DESCRIPTION_REL_ERROR_LTE, relativeError, absoluteError);
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that two values are close using a relative and/or
-   * absolute error. The relative error is symmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum relative error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String floatsGetDescriptionClose(double relativeError, float absoluteError) {
-    return floatsGetDescriptionAreClose(DESCRIPTION_REL_ERROR_LTE, relativeError, absoluteError);
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that two values are close using a relative and/or
-   * absolute error. The relative error is symmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum relative error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String longsGetDescriptionClose(double relativeError, long absoluteError) {
-    return longsGetDescriptionAreClose(DESCRIPTION_REL_ERROR_LTE, relativeError, absoluteError);
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that two values are close using a relative and/or
-   * absolute error. The relative error is symmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum relative error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @param absoluteError The maximum absolute error between <code>value1</code> and
-   *        <code>value2</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String intsGetDescriptionClose(double relativeError, int absoluteError) {
-    return intsDescriptionAreClose(DESCRIPTION_REL_ERROR_LTE, relativeError, absoluteError);
-  }
-
-  /**
-   * Get the description of the test that two values are close using a relative and absolute error.
-   * The type of relative error is specified.
-   *
-   * @param relativeErrorPrefix the relative prefix
-   * @param relativeError the relative error
-   * @param absoluteError the absolute error
-   * @return the description
-   */
-  private static String doubleGetDescriptionAreClose(String relativeErrorPrefix,
-      double relativeError, double absoluteError) {
-    // Assume both are >= 0
-
-    // Handle special cases where the test was actually equivalent to something different.
-    // 1. A relative error of 0 is the equivalent of absolute error 0, so is ignored
-    final double testRelativeError = (relativeError == 0) ? IGNORE_RELATIVE_ERROR : relativeError;
-    // 2. An absolute error of 0 is not used if the relative error is >0
-    final double testAbsoluteError =
-        (absoluteError == 0 && testRelativeError > 0) ? IGNORE_DOUBLE_ABSOLUTE_ERROR
-            : absoluteError;
-
-    // Build the description
-    final StringBuilder sb = new StringBuilder();
-    if (testRelativeError > 0) {
-      // This is always <=
-      sb.append(relativeErrorPrefix).append(testRelativeError);
-    }
-    if (testAbsoluteError >= 0) {
-      // Add combined operator
-      if (sb.length() != 0) {
-        sb.append(" || ");
-      }
-      // This may be == or <=
-      if (testAbsoluteError == 0) {
-        sb.append(DESCRIPTION_ABS_ERROR_0);
-      } else {
-        sb.append(DESCRIPTION_ABS_ERROR_LTE);
-        sb.append(testAbsoluteError);
-      }
-    }
-
-    return sb.toString();
-  }
-
-  /**
-   * Get the description of the test that two values are close using a relative and/or absolute
-   * error. The type of relative error is specified.
-   *
-   * @param relativeErrorPrefix the relative error prefix
-   * @param relativeError the relative error
-   * @param absoluteError the absolute error
-   * @return the description
-   */
-  private static String floatsGetDescriptionAreClose(String relativeErrorPrefix,
-      double relativeError, float absoluteError) {
-    // Assume both are >= 0
-
-    // Handle special cases where the test was actually equivalent to something different.
-    // 1. A relative error of 0 is the equivalent of absolute error 0, so is ignored
-    final double testRelativeError = (relativeError == 0) ? IGNORE_RELATIVE_ERROR : relativeError;
-    // 2. An absolute error of 0 is not used if the relative error is >0
-    final float testAbsoluteError =
-        (absoluteError == 0 && testRelativeError > 0) ? IGNORE_FLOAT_ABSOLUTE_ERROR : absoluteError;
-
-    // Build the description
-    final StringBuilder sb = new StringBuilder();
-    if (testRelativeError > 0) {
-      // This is always <=
-      sb.append(relativeErrorPrefix).append(testRelativeError);
-    }
-    if (testAbsoluteError >= 0) {
-      // Add combined operator
-      if (sb.length() != 0) {
-        sb.append(" || ");
-      }
-      // This may be == or <=
-      if (testAbsoluteError == 0) {
-        sb.append(DESCRIPTION_ABS_ERROR_0);
-      } else {
-        sb.append(DESCRIPTION_ABS_ERROR_LTE);
-        sb.append(testAbsoluteError);
-      }
-    }
-
-    return sb.toString();
-  }
-
-  /**
-   * Get the description of the test that two values are close using a relative and/or absolute
-   * error. The type of relative error is specified.
-   *
-   * @param relativeErrorPrefix the relative error prefix
-   * @param relativeError the relative error
-   * @param absoluteError the absolute error
-   * @return the description
-   */
-  private static String longsGetDescriptionAreClose(String relativeErrorPrefix,
-      double relativeError, long absoluteError) {
-    // Assume both are >= 0
-
-    // Handle special cases where the test was actually equivalent to something different.
-    // 1. A relative error of 0 is the equivalent of absolute error 0, so is ignored
-    final double testRelativeError = (relativeError == 0) ? IGNORE_RELATIVE_ERROR : relativeError;
-    // 2. An absolute error of 0 is not used if the relative error is >0
-    final long testAbsoluteError =
-        (absoluteError == 0 && testRelativeError > 0) ? IGNORE_LONG_ABSOLUTE_ERROR : absoluteError;
-
-    // Build the description
-    final StringBuilder sb = new StringBuilder();
-    if (testRelativeError > 0) {
-      // This is always <=
-      sb.append(relativeErrorPrefix).append(testRelativeError);
-    }
-    if (testAbsoluteError >= 0) {
-      // Add combined operator
-      if (sb.length() != 0) {
-        sb.append(" || ");
-      }
-      // This may be == or <=
-      if (testAbsoluteError == 0) {
-        sb.append(DESCRIPTION_ABS_ERROR_0);
-      } else {
-        sb.append(DESCRIPTION_ABS_ERROR_LTE);
-        sb.append(testAbsoluteError);
-      }
-    }
-
-    return sb.toString();
-  }
-
-  /**
-   * Get the description of the test that two values are close using a relative and/or absolute
-   * error. The type of relative error is specified.
-   *
-   * @param relativeErrorPrefix the relative error prefix
-   * @param relativeError the relative error
-   * @param absoluteError the absolute error
-   * @return the description
-   */
-  private static String intsDescriptionAreClose(String relativeErrorPrefix, double relativeError,
-      int absoluteError) {
-    // Assume both are >= 0
-
-    // Handle special cases where the test was actually equivalent to something different.
-    // 1. A relative error of 0 is the equivalent of absolute error 0, so is ignored
-    final double testRelativeError = (relativeError == 0) ? IGNORE_RELATIVE_ERROR : relativeError;
-    // 2. An absolute error of 0 is not used if the relative error is >0
-    final int testAbsoluteError =
-        (absoluteError == 0 && testRelativeError > 0) ? IGNORE_INT_ABSOLUTE_ERROR : absoluteError;
-
-    // Build the description
-    final StringBuilder sb = new StringBuilder();
-    if (testRelativeError > 0) {
-      // This is always <=
-      sb.append(relativeErrorPrefix).append(testRelativeError);
-    }
-    if (testAbsoluteError >= 0) {
-      // Add combined operator
-      if (sb.length() != 0) {
-        sb.append(" || ");
-      }
-      // This may be == or <=
-      if (testAbsoluteError == 0) {
-        sb.append(DESCRIPTION_ABS_ERROR_0);
-      } else {
-        sb.append(DESCRIPTION_ABS_ERROR_LTE);
-        sb.append(testAbsoluteError);
-      }
-    }
-
-    return sb.toString();
-  }
-
   //////////////////////////////////////////////////////////////////////////
   // Tests for equality using an asymmetric relative and/or absolute error.
   // Supports: double, float
@@ -1902,120 +1494,5 @@ public final class EqualityUtils {
     }
     // Compute using double precision.
     return delta <= Math.abs(expected) * relativeError;
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that a value is close to an expected value using a relative
-   * and absolute error. The relative error is asymmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum error, relative to <code>expected</code>, between
-   *        <code>expected</code> and <code>actual</code> for which both numbers are still
-   *        considered equal.
-   * @param absoluteError The maximum absolute error between <code>expected</code> and
-   *        <code>actual</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String doublesGetDescriptionIsCloseTo(double relativeError, double absoluteError) {
-    return doubleGetDescriptionAreClose(DESCRIPTION_ASYM_REL_ERROR_LTE, relativeError,
-        absoluteError);
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that a value is close to an expected value using a relative
-   * and absolute error. The relative error is asymmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum error, relative to <code>expected</code>, between
-   *        <code>expected</code> and <code>actual</code> for which both numbers are still
-   *        considered equal.
-   * @param absoluteError The maximum absolute error between <code>expected</code> and
-   *        <code>actual</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String floatsGetDescriptionIsCloseTo(double relativeError, float absoluteError) {
-    return floatsGetDescriptionAreClose(DESCRIPTION_ASYM_REL_ERROR_LTE, relativeError,
-        absoluteError);
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that a value is close to an expected value using a relative
-   * and absolute error. The relative error is asymmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum error, relative to <code>expected</code>, between
-   *        <code>expected</code> and <code>actual</code> for which both numbers are still
-   *        considered equal.
-   * @param absoluteError The maximum absolute error between <code>expected</code> and
-   *        <code>actual</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String longsGetDescriptionIsCloseTo(double relativeError, long absoluteError) {
-    return longsGetDescriptionAreClose(DESCRIPTION_ASYM_REL_ERROR_LTE, relativeError,
-        absoluteError);
-  }
-
-  //@formatter:off
-  /**
-   * Get the description of the test that a value is close to an expected value using a relative
-   * and absolute error. The relative error is asymmetric for {@code value1} and {@code value2}.
-   *
-   * <p>It is assumed the errors have been validated (are positive and finite).
-   *
-   * <p>Note the special cases:
-   *
-   * <ul>
-   * <li>When the relative error is zero this is equivalent to an absolute error of zero.
-   * The resulting description just contains the absolute error.
-   * <li>When the absolute error is zero and the relative error is above zero the resulting
-   * description just contains the relative error.
-   * </ul>
-   *
-   * @param relativeError The maximum error, relative to <code>expected</code>, between
-   *        <code>expected</code> and <code>actual</code> for which both numbers are still
-   *        considered equal.
-   * @param absoluteError The maximum absolute error between <code>expected</code> and
-   *        <code>actual</code> for which both numbers are still considered equal.
-   * @return the description
-   */
-  //@formatter:on
-  static String intsGetDescriptionIsCloseTo(double relativeError, int absoluteError) {
-    return intsDescriptionAreClose(DESCRIPTION_ASYM_REL_ERROR_LTE, relativeError, absoluteError);
   }
 }
