+++ gdsc-test-api/src/main/resources/uk/ac/sussex/gdsc/test/api/comparisons/NumberNumberClose.properties
@@ -21,6 +21,5 @@
 # <http://www.gnu.org/licenses/gpl-3.0.html>.
 # #L%
 ###
-classname.Float = Double Float
-class.relError = 1e-3 1e-3
-class.absError = 1e-6 1e-6f
+classname.Number = Byte Double Float Int Long Short
+class.errorType = int double float long long int
