+++ gdsc-test-api/src/main/resources/uk/ac/sussex/gdsc/test/api/comparisons/NumberNumberIsCloseTo.st
@@ -1,22 +1,22 @@
 <package>
 
-import uk.ac.sussex.gdsc.test.api.function.<Float><Float>BiPredicate;
+import uk.ac.sussex.gdsc.test.api.function.<Number><Number>BiPredicate;
 
 import java.io.Serializable;
 import java.util.function.Supplier;
 
 /**
- * Tests an {@code actual} {@code <float>} value is close to an {@code expected} value using a
+ * Tests an {@code actual} {@code <number>} value is close to an {@code expected} value using a
  * relative and/or an absolute error tolerance.
  *
  * \<p>The computed relative error is asymmetric and the test is equivalent to testing testing
  * {@code actual} falls within a relative and/or absolute range of {@code expected}.
  *
- * @see <Float><Float>BiPredicate#test(<float>, <float>)
- * @see EqualityUtils#areClose(<float>, <float>, double, <float>)
+ * @see <Number><Number>BiPredicate#test(<number>, <number>)
+ * @see EqualityUtils#areClose(<number>, <number>, double, <number>)
  */
-public class <Float><Float>IsCloseTo
-    implements <Float><Float>BiPredicate, Supplier\<String>, Serializable {
+public class <Number><Number>IsCloseTo
+    implements <Number><Number>BiPredicate, Supplier\<String>, Serializable {
 
   /** The serial version ID. */
   private static final long serialVersionUID = 1L;
@@ -24,10 +24,10 @@ public class <Float><Float>IsCloseTo
   /** The relative error. */
   private final double relativeError;
   /** The absolute error. */
-  private final <float> absoluteError;
+  private final <errorType> absoluteError;
 
   /**
-   * Create a predicate to test a {@code <float>} value is close to an expected value using a
+   * Create a predicate to test a {@code <number>} value is close to an expected value using a
    * relative and/or absolute error. The relative error between values {@code expected} and
    * {@code actual} is relative to the magnitude of {@code expected} and the test is:
    *
@@ -45,10 +45,11 @@ public class <Float><Float>IsCloseTo
    * @param absoluteError The maximum absolute error between \<code>expected\</code> and
    *        \<code>actual\</code> for which both numbers are still considered equal.
    * @throws IllegalArgumentException If the relative error is not positive finite
-   * @throws IllegalArgumentException If the absolute error is not positive finite
+   * @throws IllegalArgumentException If the absolute error is not positive or is
+   *         <code>>=</code> than the maximum difference between <number> primitives
    */
-  public <Float><Float>IsCloseTo(double relativeError, <float> absoluteError) {
-    EqualityUtils.<float>sValidateIsCloseTo(relativeError, absoluteError);
+  public <Number><Number>IsCloseTo(double relativeError, <errorType> absoluteError) {
+    EqualityUtils.<number>sValidateIsCloseTo(relativeError, absoluteError);
     this.relativeError = relativeError;
     this.absoluteError = absoluteError;
   }
@@ -62,9 +63,9 @@ public class <Float><Float>IsCloseTo
    * @return true if actual is close to expected
    */
   @Override
-  public boolean test(<float> expected, <float> actual) {
+  public boolean test(<number> expected, <number> actual) {
     // Errors have been checked so use the package level method
-    return EqualityUtils.<float>sTestIsCloseTo(expected, actual, relativeError, absoluteError);
+    return EqualityUtils.<number>sTestIsCloseTo(expected, actual, relativeError, absoluteError);
   }
 
   /**
@@ -74,6 +75,6 @@ public class <Float><Float>IsCloseTo
    */
   @Override
   public String get() {
-    return EqualityUtils.<float>sGetDescriptionIsCloseTo(relativeError, absoluteError);
+    return DescriptionHelper.getDescriptionIsCloseTo(relativeError, absoluteError);
   }
 }
