+++ gdsc-test-api/src/main/resources/uk/ac/sussex/gdsc/test/api/comparisons/NumberNumberWithin.properties
@@ -24,4 +24,5 @@
 classname.Number = Byte Double Float Int Long Long Short
 classname.Within = Within Within Within Within Within BigWithin Within
 class.errorType = int double float long long BigInteger int
+class.descriptionType = int double float long long long int
 class.import1 = [] [] [] [] [] java.math.BigInteger []
