+++ gdsc-test-api/src/test/resources/uk/ac/sussex/gdsc/test/api/comparisons/NumberNumberCloseTest.properties
@@ -21,4 +21,6 @@
 # <http://www.gnu.org/licenses/gpl-3.0.html>.
 # #L%
 ###
-classname.Float = Double Float
+classname.Number = Byte Double Float Int Long Short
+class.errorType = int double float long long int
+class.absError = 2 1e-6 1e-6f 2 2 2
